+++ tests/fixes/test_funcs.py
@@ -13,10 +13,17 @@ from pytestify.fixes.funcs import rewrite_pytest_funcs
         ),
         ('self.assertWarns(Exc)', 'pytest.warns(Exc)'),
         ('with self.assertWarns(Exc): pass', 'with pytest.warns(Exc): pass'),
+        ('self.fail("some reason")', 'pytest.fail("some reason")'),
+        (
+            '@unittest.expectedFailure\n'
+            'def blah(): pass',
+            '@pytest.mark.xfail\n'
+            'def blah(): pass',
+        ),
         (
             '@unittest.skip("some reason")\n'
             'def blah(): pass',
-            '@pytest.skip("some reason")\n'
+            '@pytest.mark.skip("some reason")\n'
             'def blah(): pass',
         ),
         (
@@ -25,6 +32,7 @@ from pytestify.fixes.funcs import rewrite_pytest_funcs
             '@pytest.mark.skipif(some_bool)\n'
             'def blah(): pass',
         ),
+        ('unittest.skipTest("Some reason")', 'pytest.skip("Some reason")'),
         (
             '@unittest.skipUnless(some_bool)\n'
             'def blah(): pass',
@@ -80,3 +88,15 @@ def test_adds_import_when_necessary(before, after):
 )
 def test_doesnt_rewrite_pytest_funcs(line):
     assert rewrite_pytest_funcs(line) == line
+
+
+@pytest.mark.parametrize(
+    'line', [
+        'unittest.expectedFailure',
+        'unittest.skipIf(some_bool)',
+        'unittest.skipUnless(some_bool)',
+    ],
+)
+def test_syntax_error_on_bad_funcs(line):
+    with pytest.raises(SyntaxError):
+        rewrite_pytest_funcs(line)